requirements.txt
fastapi
uvicorn
pydantic
click
python-multipart
python-dotenv